Split parser tests to add new parsers
