+++ multibody/multibody_tree/parsing/test/multibody_plant_parser_common_test.cc
@@ -1,0 +1,156 @@
+#include <string>
+
+#include <gtest/gtest.h>
+
+#include "drake/geometry/scene_graph.h"
+#include "drake/multibody/multibody_tree/multibody_plant/multibody_plant.h"
+#include "drake/multibody/multibody_tree/parsing/parser_common.h"
+#include "drake/multibody/multibody_tree/parsing/test/multibody_plant_parser_test_loaders.h"
+#include "drake/systems/framework/context.h"
+
+namespace drake {
+namespace multibody {
+namespace parsing {
+namespace {
+
+using geometry::GeometryId;
+using multibody_plant::CoulombFriction;
+using multibody_plant::MultibodyPlant;
+
+// Fixture to setup a simple model with both collision and visual geometry,
+// loaded with the SDF parser.
+class MultibodyPlantLinkTests :
+      public testing::TestWithParam<test::ModelLoadFunction> {
+ public:
+  void SetUp() override {
+    base_name_ = "drake/multibody/multibody_tree/parsing/test/"
+        "links_with_visuals_and_collisions";
+  }
+
+  // Loads the MultibodyPlant part of the model. Geometry is ignored.
+  void LoadMultibodyPlantOnly() {
+    auto load_model = GetParam();
+    load_model(base_name_, &plant_, nullptr);
+    plant_.Finalize();
+  }
+
+  // Loads the entire model including the multibody dynamics part of it and the
+  // geometries for both visualization and contact modeling.
+  void LoadMultibodyPlantAndSceneGraph() {
+    auto load_model = GetParam();
+    load_model(base_name_, &plant_, &scene_graph_);
+    plant_.Finalize(&scene_graph_);
+  }
+
+ protected:
+  MultibodyPlant<double> plant_;
+  geometry::SceneGraph<double> scene_graph_;
+  std::string base_name_;
+};
+
+TEST_P(MultibodyPlantLinkTests, LinkWithVisuals) {
+  LoadMultibodyPlantAndSceneGraph();
+
+  EXPECT_EQ(plant_.num_bodies(), 4);  // It includes the world body.
+  EXPECT_EQ(plant_.num_visual_geometries(), 5);
+
+  const std::vector<GeometryId>& link1_visual_geometry_ids =
+      plant_.GetVisualGeometriesForBody(plant_.GetBodyByName("link1"));
+  EXPECT_EQ(link1_visual_geometry_ids.size(), 2);
+
+  const std::vector<GeometryId>& link2_visual_geometry_ids =
+      plant_.GetVisualGeometriesForBody(plant_.GetBodyByName("link2"));
+  EXPECT_EQ(link2_visual_geometry_ids.size(), 3);
+
+  const std::vector<GeometryId>& link3_visual_geometry_ids =
+      plant_.GetVisualGeometriesForBody(plant_.GetBodyByName("link3"));
+  EXPECT_EQ(link3_visual_geometry_ids.size(), 0);
+
+  // TODO(SeanCurtis-TRI): Once SG supports it, confirm `GeometryId` maps to the
+  // correct geometry.
+}
+
+// Verifies we can still parse the model dynamics if a SceneGraph is not
+// supplied.
+TEST_P(MultibodyPlantLinkTests, ParseWithoutASceneGraph) {
+  LoadMultibodyPlantOnly();
+
+  EXPECT_EQ(plant_.num_bodies(), 4);  // It includes the world body.
+  EXPECT_EQ(plant_.num_visual_geometries(), 0);
+}
+
+// Verifies that the source registration with a SceneGraph can happen before a
+// call to AddModelFromSdfFile().
+TEST_P(MultibodyPlantLinkTests, RegisterWithASceneGraphBeforeParsing) {
+  plant_.RegisterAsSourceForSceneGraph(&scene_graph_);
+  LoadMultibodyPlantAndSceneGraph();
+
+  EXPECT_EQ(plant_.num_bodies(), 4);  // It includes the world body.
+  EXPECT_EQ(plant_.num_visual_geometries(), 5);
+
+  using geometry::GeometryId;
+  const std::vector<GeometryId>& link1_visual_geometry_ids =
+      plant_.GetVisualGeometriesForBody(plant_.GetBodyByName("link1"));
+  EXPECT_EQ(link1_visual_geometry_ids.size(), 2);
+
+  // TODO(sam.creasey) Verify that the path to the mesh for the second
+  // visual geometry on link 1 is resolved correctly.  Currently the
+  // resolved mesh filename is trapped inside the shape object within
+  // the scene graph and I can't find any good way to dig it back out.
+  // It would be possible to modify geometry::DispatchLoadMessage to
+  // take a DrakeLcmInterface and then scrape the filename out of the
+  // resulting lcmt_viewer_load_robot message, but I don't want to.
+
+  const std::vector<GeometryId>& link2_visual_geometry_ids =
+      plant_.GetVisualGeometriesForBody(plant_.GetBodyByName("link2"));
+  EXPECT_EQ(link2_visual_geometry_ids.size(), 3);
+
+  const std::vector<GeometryId>& link3_visual_geometry_ids =
+      plant_.GetVisualGeometriesForBody(plant_.GetBodyByName("link3"));
+  EXPECT_EQ(link3_visual_geometry_ids.size(), 0);
+
+  // TODO(SeanCurtis-TRI): Once SG supports it, confirm `GeometryId` maps to the
+  // correct geometry.
+}
+
+// Verifies we can parse link collision geometries and surface friction.
+TEST_P(MultibodyPlantLinkTests, LinksWithCollisions) {
+  LoadMultibodyPlantAndSceneGraph();
+
+  EXPECT_EQ(plant_.num_bodies(), 4);  // It includes the world body.
+  EXPECT_EQ(plant_.num_collision_geometries(), 3);
+
+  using geometry::GeometryId;
+  const std::vector<GeometryId>& link1_collision_geometry_ids =
+      plant_.GetCollisionGeometriesForBody(plant_.GetBodyByName("link1"));
+  ASSERT_EQ(link1_collision_geometry_ids.size(), 2);
+
+  EXPECT_TRUE(
+      plant_.default_coulomb_friction(link1_collision_geometry_ids[0]) ==
+          CoulombFriction<double>(0.8, 0.3));
+  EXPECT_TRUE(
+      plant_.default_coulomb_friction(link1_collision_geometry_ids[1]) ==
+          CoulombFriction<double>(1.5, 0.6));
+
+  const std::vector<GeometryId>& link2_collision_geometry_ids =
+      plant_.GetCollisionGeometriesForBody(plant_.GetBodyByName("link2"));
+  ASSERT_EQ(link2_collision_geometry_ids.size(), 0);
+
+  const std::vector<GeometryId>& link3_collision_geometry_ids =
+      plant_.GetCollisionGeometriesForBody(plant_.GetBodyByName("link3"));
+  ASSERT_EQ(link3_collision_geometry_ids.size(), 1);
+  // Verifies the default value of the friction coefficients when the user does
+  // not specify them in the SDF file.
+  EXPECT_TRUE(
+      plant_.default_coulomb_friction(link3_collision_geometry_ids[0]) ==
+      parsing::default_friction());
+}
+
+
+INSTANTIATE_TEST_CASE_P(SdfMultibodyPlantLinkTests,
+                        MultibodyPlantLinkTests,
+                        ::testing::Values(test::LoadFromSdf));
+}  // namespace
+}  // namespace parsing
+}  // namespace multibody
+}  // namespace drake
